% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{combined_rmse}
\alias{combined_rmse}
\title{Calculate the combined root mean square error}
\usage{
combined_rmse(breakpoint, data, tau = 0.05)
}
\arguments{
\item{breakpoint}{duration to segment the dive data}

\item{data}{dive data with breaths and dives identified (see
\code{\link[=find_breaths_dives]{find_breaths_dives()}})}

\item{tau}{which quantile to regress (0.05 by default)}
}
\value{
the combined RMSE of the segmented regression
}
\description{
\code{\link[=combined_rmse]{combined_rmse()}} segments dive data by duration, fits two quantile
regressions, and returns the combined root mean square error (RMSE) of both
regressions.
}
