% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{decimate}
\alias{decimate}
\title{Decimate a record}
\usage{
decimate(x, new_fs)
}
\arguments{
\item{x}{a depth/time record}

\item{new_fs}{sampling frequency to decimate to}
}
\value{
a record with the new sampling frequency
}
\description{
Decimate a record
}
