% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dn_to_posix}
\alias{dn_to_posix}
\title{Convert MATLAB date numbers to POSIXct}
\usage{
dn_to_posix(dn, tz = "UTC")
}
\arguments{
\item{dn}{Vector of MATLAB date numbers (numeric)}

\item{tz}{Time zone (character)}
}
\value{
POSIXct vector
}
\description{
Convert MATLAB date numbers to POSIXct
}
\examples{

# "2020-01-27 19:12:00 UTC"
dn_to_posix(737817.8)
}
