% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_breaths.R
\name{find_breaths}
\alias{find_breaths}
\title{Breath classifier}
\usage{
find_breaths(prh, period, prominence, surf_thr)
}
\arguments{
\item{prh}{\verb{[data.frame]} with column p2 (smoothed depth)}

\item{period}{minimum interval between breaths}

\item{prominence}{minimum depth prominence of breath}
}
\value{
\verb{[data.frame]} prh filtered to breaths with surface_idx and row_idx
}
\description{
Breath classifier
}
\examples{
breaths <- find_breaths(feat, period = 5, prominence = 0.5, surf_thr = 0.5)
}
