% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find_breaths_dives}
\alias{find_breaths_dives}
\title{Find breaths and dives in depth data}
\usage{
find_breaths_dives(x, interval, surface, ibi_thr)
}
\arguments{
\item{x}{a depth/time object (see \code{\link[=read_depth]{read_depth()}})}

\item{interval}{typical interval between breaths in seconds (e.g. 10 for blue whales)}

\item{surface}{minimum depth for a breath}

\item{ibi_thr}{inter-breath interval threshold for a dive}
}
\value{
a depth/time object with identified breaths and dives
}
\description{
Find breaths and dives in depth data
}
