% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prh.R
\name{find_nc}
\alias{find_nc}
\title{Find a PRH in NetCDF format}
\usage{
find_nc(depid, cats_path)
}
\arguments{
\item{depid}{deployment identifier e.g. bw180827-46}

\item{cats_path}{location of CATS folder}
}
\value{
absolute path to NetCDF PRH file
}
\description{
Find a PRH in NetCDF format
}
\examples{
find_nc("bw180827-46", "/Volumes/COPYCATSdat/CATS/")
}
