% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_breaths.R
\name{find_peaks}
\alias{find_peaks}
\title{Robust peak finder}
\usage{
find_peaks(x, width, prominence, idx_off = 0)
}
\arguments{
\item{x}{\link{vector}}

\item{width}{\link{vector(1)} width of search window}

\item{prominence}{\link{vector(1)} minimum prominence of peaks}

\item{idx_off}{\link{vector(1)} internal use only}
}
\value{
\link{matrix(n, 2)} first column is peak index, second column is peak
prominence. Ordered by prominence.
}
\description{
Robust peak finder
}
\examples{
set.seed(1202)
x <- seq(0, 4 * pi, length.out = 1e2)
y <- sin(x^1.4) + rnorm(length(x), sd = 0.25)
peaks <- find_peaks(y, width = 10, prominence = 0.5)
plot(seq_along(x), y, type = "l")
abline(v = peaks[,1], lty = 3)
}
