% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_adl}
\alias{fit_adl}
\title{Fit a behavioral aerobic dive limit model to dive data}
\usage{
fit_adl(x, n_dives = 1, nofeeding = TRUE)
}
\arguments{
\item{x}{dive data with dives/breaths identified (see \code{\link[=find_breaths_dives]{find_breaths_dives()}})}

\item{n_dives}{number of dives to average durations and breaths over}

\item{nofeeding}{if true, removes dives with lunges}
}
\value{
a list with the aerobic dive limit (ADL), quantile regressions left and right of the ADL, whether it passes the continuity requirement, the original dive data, and the n_dives and nofeeding parameter values.
}
\description{
Roughly following the methods of Horning (2012), \code{\link[=fit_adl]{fit_adl()}} searches for a duration beyond which the post-dive recovery increases.
}
\examples{

fit_adl(bm181021_dives) \%>\%
  plot_adl()
}
