% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prh.R
\name{get_depid}
\alias{get_depid}
\alias{get_fs}
\alias{get_tzone.prh}
\title{PRH metadata accessors}
\usage{
get_depid(prh)

get_fs(prh)

\method{get_tzone}{prh}(prh)
}
\arguments{
\item{prh}{a PRH object}
}
\value{
PRH metadata e.g. the deployment identifier (\code{get_depid()}) or timezone (\code{get_tzone()})
}
\description{
PRH metadata accessors
}
\examples{
bw180905_53 <- read_prh_nc("bw180905-53", "/Volumes/COPYCATSdat/CATS/")
get_depid(bw180905_53)
get_tzone(bw180905_53)
}
