% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_adl}
\alias{plot_adl}
\title{Plot an aerobic dive limit model}
\usage{
plot_adl(fitted_adl, ylim = c(0, 25))
}
\arguments{
\item{fitted_adl}{an ADL model (see \code{\link[=fit_adl]{fit_adl()}})}

\item{ylim}{number of breaths to limit the y-axes (0-25 by default)}
}
\value{
a ggplot object with post-dive breath counts plotted against dive duration superimposed with the ADL and the intersecting quantile regression lines.
}
\description{
After fitting an aerobic dive limit (ADL) model using \code{\link[=fit_adl]{fit_adl()}}, \code{\link[=plot_adl]{plot_adl()}} visualizes the ADL, the quantile regressions, and their intersection.
}
\examples{

fit_adl(bm181021_dives) \%>\%
  plot_adl()
}
