% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_dive_event}
\alias{plot_dive_event}
\title{Plot a dive event (dive + post-dive interval)}
\usage{
plot_dive_event(x, diveid, buffer)
}
\arguments{
\item{x}{dive data with dives/breaths identified (see \code{\link[=find_breaths_dives]{find_breaths_dives()}})}

\item{diveid}{dive to plot}

\item{buffer}{time around dive to plot (in seconds)}
}
\value{
a ggplot showing depth over time with breaths indicated in red
}
\description{
Plot a dive event (dive + post-dive interval)
}
