% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_breaths.R
\name{plot_surface}
\alias{plot_surface}
\title{Plot breaths in a surface interval}
\usage{
plot_surface(prh, surface_idx)
}
\arguments{
\item{prh}{\verb{[data.frame]} prh with breaths identified}

\item{surface_idx}{\verb{[integer(1)]}}
}
\value{
\verb{[grob]}
}
\description{
Plot breaths in a surface interval
}
\examples{
breaths <- find_breaths(feat, period = 5, prominence = 0.5, surf_thr = 0.5)
plot_surface(breaths, 1)
}
