% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_deployment}
\alias{read_deployment}
\title{Read deployment from ncdf file}
\usage{
read_deployment(path, tz = "UTC")
}
\arguments{
\item{path}{File path to ncdf file}

\item{tz}{Time zone of deployment}
}
\value{
list with depth-time tibble (dt, p), vector of lunge times, and
sampling rate (fs)
}
\description{
Read deployment from ncdf file
}
