% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_depth}
\alias{read_depth}
\title{Read depth/time record from ncdf file}
\usage{
read_depth(path, tz = "UTC")
}
\arguments{
\item{path}{File path to ncdf file}

\item{tz}{Time zone of deployment}
}
\value{
list with depth-time tibble (dt, p) and sampling rate (fs)
}
\description{
Read depth/time record from ncdf file
}
