% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prh.R
\name{read_prh_nc}
\alias{read_prh_nc}
\title{Read a PRH from NetCDF format}
\usage{
read_prh_nc(depid, cats_path)
}
\arguments{
\item{depid}{deployment identifier e.g. bw180827-46}

\item{cats_path}{location of CATS folder}
}
\value{
a PRH object
}
\description{
Read a PRH from NetCDF format
}
\examples{
bw180905_53 <- read_prh_nc("bw180905-53", "/Volumes/COPYCATSdat/CATS/")
}
