% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_dives}
\alias{summarize_dives}
\title{Summarize dives in a deployment}
\usage{
summarize_dives(x)
}
\arguments{
\item{x}{depth/time object with breaths and dives identified}
}
\value{
a tibble summarizing each dive:
\itemize{
\item{duration}
\item{depth}
\item{n_breaths}
}
}
\description{
Summarize dives in a deployment
}
